tasks/prepare_opa_assessments/requirements.txt
functions_framework==3.8.2
google-cloud-storage==3.1.0
python-dotenv==1.0.1
